+++ mobile/src/navigation/MatchesStackNavigator.tsx
@@ -1,7 +1,5 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import { MatchDetailScreen } from '../screens/MatchDetailScreen';
-import { MatchesScreen } from '../screens/MatchesScreen';
 import { matchDetailScreenOptions } from './matchDetailScreenOptions';
 import { screenSubtitles, screenTitles } from './screenTitles';
 import { StackHeaderTitle } from './StackHeaderTitle';
@@ -15,7 +13,7 @@ export function MatchesStackNavigator() {
     <Stack.Navigator screenOptions={stackScreenOptions}>
       <Stack.Screen
         name="Matches"
-        component={MatchesScreen}
+        getComponent={() => require('../screens/MatchesScreen').MatchesScreen}
         options={{
           headerTitle: () => (
             <StackHeaderTitle
@@ -27,7 +25,7 @@ export function MatchesStackNavigator() {
       />
       <Stack.Screen
         name="MatchDetail"
-        component={MatchDetailScreen}
+        getComponent={() => require('../screens/MatchDetailScreen').MatchDetailScreen}
         options={matchDetailScreenOptions}
       />
     </Stack.Navigator>